fix: restaurar login Google com popup no PWA iOS
Reverte o fluxo de autenticação quebrado (AuthProvider, redirect e
signInWithRedirect no PWA) para o padrão original com signInWithPopup
e onAuthStateChanged.

- Remove AuthContext, AuthRedirect e googleAuth helper
- Aguarda authStateReady antes de redirecionar após login Google
- Evita race condition no guard de /home com authStateReady
- Carrega SignIn, Register e HomeAuth sem lazy loading
- Restaura getFirestore simples no Firebase
- Força atualização do service worker no PWA
