--- src/context/AuthContext.jsx
@@ -1,53 +1,0 @@
-import { createContext, useContext, useEffect, useMemo, useState } from "react";
-import { onAuthStateChanged } from "firebase/auth";
-import { auth } from "../../config/firebase";
-import { resolveGoogleRedirectResult } from "../utils/googleAuth";
-
-const AuthContext = createContext({
-  user: null,
-  loading: true,
-});
-
-export function AuthProvider({ children }) {
-  const [user, setUser] = useState(null);
-  const [loading, setLoading] = useState(true);
-
-  useEffect(() => {
-    let unsubscribe = () => {};
-
-    const initAuth = async () => {
-      try {
-        await resolveGoogleRedirectResult();
-        await auth.authStateReady();
-        setUser(auth.currentUser);
-      } catch (error) {
-        console.error("Auth initialization failed:", error);
-      } finally {
-        setLoading(false);
-      }
-
-      unsubscribe = onAuthStateChanged(auth, (nextUser) => {
-        setUser(nextUser);
-        setLoading(false);
-      });
-    };
-
-    initAuth();
-
-    return () => unsubscribe();
-  }, []);
-
-  const value = useMemo(
-    () => ({
-      user,
-      loading,
-    }),
-    [user, loading]
-  );
-
-  return <AuthContext.Provider value={value}>{children}</AuthContext.Provider>;
-}
-
-export function useAuth() {
-  return useContext(AuthContext);
-}
