+++ src/pages/SignIn/index.jsx
@@ -1,12 +1,11 @@
 import { Link } from "react-router-dom";
-import styles from "../Register/styles.module.scss"; // Certifique-se de que os estilos estão corretos
-import logo from "../../assets/WalletIcon.png"; // Caminho para o logo
-import { useState } from "react";
-import { auth } from "../../../config/firebase";
+import styles from "../Register/styles.module.scss";
+import logo from "../../assets/WalletIcon.png";
+import { useEffect, useState } from "react";
+import { auth, googleProvider } from "../../../config/firebase";
 import { useNavigate } from "react-router-dom";
-import { signInWithEmailAndPassword } from "firebase/auth";
+import { signInWithEmailAndPassword, signInWithPopup } from "firebase/auth";
 import { FaEye, FaEyeSlash } from "react-icons/fa";
-import { signInWithGoogle } from "../../utils/googleAuth";
 
 export default function SignIn() {
   const navigate = useNavigate();
@@ -14,6 +13,15 @@ export default function SignIn() {
   const [password, setPassword] = useState("");
   const [showPassword, setShowPassword] = useState(false);
   const [errorMessage, setErrorMessage] = useState("");
+
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged((user) => {
+      if (user) {
+        navigate("/home/expenses", { replace: true });
+      }
+    });
+    return () => unsubscribe();
+  }, [navigate]);
 
   const toggleShowPassword = () => {
     setShowPassword(prevShowPassword => !prevShowPassword);
@@ -32,13 +40,13 @@ export default function SignIn() {
 
   const continueWithGoogle = async () => {
     try {
-      const result = await signInWithGoogle();
-      if (result?.user) {
-        navigate("/home/expenses");
-      }
+      await signInWithPopup(auth, googleProvider);
+      await auth.authStateReady();
+      navigate("/home/expenses", { replace: true });
       setErrorMessage("");
     } catch (err) {
       console.error("Error with Google sign-in:", err.message);
+      setErrorMessage("Google sign-in failed. Please try again.");
     }
   };
