+++ src/pages/homeAuth/index.jsx
@@ -2,17 +2,14 @@ import walletIcon from "../../assets/WalletIcon.png";
 import styles from "./styles.module.scss";
 import { auth } from "../../../config/firebase";
 import { Link, Outlet, useNavigate } from "react-router-dom";
-import { signOut } from "firebase/auth";
+import { onAuthStateChanged, signOut } from "firebase/auth";
 import x from "../../assets/x.svg";
 import { useEffect, useState } from "react";
 import menu from "../../assets/menu.svg";
 import { useBodyScrollLock } from "../../hooks/useBodyScrollLock";
-import LoadingComponent from "../../components/LoadingComponent/LoadingComponent";
-import { useAuth } from "../../context/AuthContext";
 
 function HomeAuth() {
   const navigate = useNavigate();
-  const { user, loading } = useAuth();
   const [menuOpen, setMenuOpen] = useState(false);
   const [isMobile, setIsMobile] = useState(window.innerWidth <= 768);
 
@@ -39,23 +36,32 @@ function HomeAuth() {
       }
     };
 
+    let unsubscribe = () => {};
+
+    const setupAuthGuard = async () => {
+      await auth.authStateReady();
+      unsubscribe = onAuthStateChanged(auth, (user) => {
+        if (!user) {
+          navigate("/signin", { replace: true });
+        }
+      });
+    };
+
+    setupAuthGuard();
     handleResize();
     window.addEventListener("resize", handleResize);
 
     return () => {
       window.removeEventListener("resize", handleResize);
+      unsubscribe();
     };
-  }, []);
+  }, [navigate]);
 
   const handleNavClick = () => {
     if (isMobile) {
       closeMenu();
     }
   };
-
-  if (loading || !user) {
-    return <LoadingComponent />;
-  }
 
   const menuContent = (
     <>
